@@ -9,7 +9,7 @@
  * @author CleverMonitor <support@clevermonitor.com>
  */
 
-require __DIR__ . '/vendor/autoload.php';
+require __DIR__ . '/../vendor/autoload.php';
 
 $cmApi = new \CleverMonitor\Api\Methods\Lists('Your ID', 'Your Token');
 
@@ -17,33 +17,33 @@ $cmApi = new \CleverMonitor\Api\Methods\Lists('Your ID', 'Your Token');
  * Create a distribution list
  * @POST
  */
-$data = array(
+$data = [
 	'name' => 'List',
-	'emails' => array(
-		0 => array(
+	'emails' => [
+		0 => [
 			'email' => 'john.doe@example.com',
 			'first_name' => 'John',
 			'last_name' => 'Doe'
-		)
-	)
-);
+		]
+	]
+];
 $create = $cmApi->postList($data);
 
 $httpResponseCode = $create->getStatusCode();
 $httpResponseData = $create->getData();
 
 /**
- * Overview if distribution list
+ * Overview of a distribution lists
  * @GET
  */
 
-$report = $cmApi->getLists(100, 0);
+$overview = $cmApi->getLists(100, 0);
 
-$httpResponseCode = $create->getStatusCode();
-$httpResponseData = $report->getData();
+$httpResponseCode = $overview->getStatusCode();
+$httpResponseData = $overview->getData();
 
 /**
- * Detail of distribution list
+ * Detail of a distribution list
  * @GET
  */
 $detail = $cmApi->getList('5ff42ca6-1965-49ed-97d0-b2b568c88bfd');
@@ -55,9 +55,9 @@ $httpResponseData = $detail->getData();
  * Update a distribution list
  * @PATCH
  */
-$data = array(
+$data = [
 	'name' => 'List Group'
-	);
+];
 $update = $cmApi->patchList('5ff42ca6-1965-49ed-97d0-b2b568c88bfd', $data);
 
 $httpResponseCode = $update->getStatusCode();
@@ -85,15 +85,15 @@ $httpResponseData = $preview->getData();
  * Adding subscribers to a distribution list
  * @PUT
  */
-$data = array(
-	'emails' => array(
-		0 => array(
+$data = [
+	'emails' => [
+		0 => [
 			'email' => 'john.doe@example.com',
 			'first_name' => 'John',
 			'last_name' => 'Doe'
-		)
-	)
-);
+		]
+	]
+];
 $add = $cmApi->putListSubscribers('5ff42ca6-1965-49ed-97d0-b2b568c88bfd', $data);
 
 $httpResponseCode = $add->getStatusCode();
@@ -103,13 +103,13 @@ $httpResponseData = $add->getData();
  * Delete subscribers in a distribution list
  * @DELETE
  */
-$data = array(
-	'emails' => array(
-		0 => array(
+$data = [
+	'emails' => [
+		0 => [
 			'email' => 'john.doe@example.com',
-		)
-	)
-);
+		]
+	]
+];
 $delete = $cmApi->deleteListSubscribers('5ff42ca6-1965-49ed-97d0-b2b568c88bfd', $data);
 
 $httpResponseCode = $delete->getStatusCode();
