@@ -4,18 +4,16 @@
 	/**
 	 * CleverMonitor Developers
 	 * 
-	 * Lists
+	 * Distribution Lists
 	 * @author CleverMonitor <support@clevermonitor.com>
 	 */
 
 	class Lists extends Base {
 
 		/**
-		 * CleverMonitor Developers
-		 *
 		 * Create a distribution list
 		 * @param array $data
-		 * @return \stdClass
+		 * @return \CleverMonitor\Api\Connection\Response
 		 */
 		public function postList(array $data) {
 			$curl = $this->connection->post('lists/', $data);
@@ -23,24 +21,24 @@
 		}
 
 		/**
-		 * Overview a distribution list
+		 * Overview of a distribution lists
 		 * @param int $count
 		 * @param int $offset
-		 * @return \stdClass
+		 * @return \CleverMonitor\Api\Connection\Response
 		 */
 		public function getLists($count = 100, $offset = 0) {
-			$query = array(
+			$query = [
 				'count' => $count,
 				'offset' => $offset,
-			);
+			];
 			$curl = $this->connection->get('lists/', $query);
 			return $curl;
 		}
 
 		/**
-		 *  Detail of distribution list
+		 * Detail of a distribution list
 		 * @param string $list_id
-		 * @return \stdClass
+		 * @return \CleverMonitor\Api\Connection\Response
 		 */
 		public function getList($list_id) {
 			$curl = $this->connection->get('lists/' . $list_id);
@@ -51,7 +49,7 @@
 		 * Update a distribution list
 		 * @param string $list_id
 		 * @param array $data
-		 * @return \stdClass
+		 * @return \CleverMonitor\Api\Connection\Response
 		 */
 		public function patchList($list_id, array $data) {
 			$curl = $this->connection->patch('lists/' . $list_id, $data);
@@ -61,7 +59,7 @@
 		/**
 		 * Delete a distribution list
 		 * @param string $list_id
-		 * @return \stdClass
+		 * @return \CleverMonitor\Api\Connection\Response
 		 */
 		public function deleteList($list_id) {
 			$curl = $this->connection->delete('lists/' . $list_id);
@@ -73,13 +71,13 @@
 		 * @param string $list_id
 		 * @param int $count
 		 * @param int $offset
-		 * @return \stdClass
+		 * @return \CleverMonitor\Api\Connection\Response
 		 */
 		public function getListSubscribers($list_id, $count = 100, $offset = 0) {
-			$query = array(
+			$query = [
 				'count' => $count,
 				'offset' => $offset,
-			);
+			];
 			$curl = $this->connection->get('lists/'. $list_id . '/subscribers/', $query);
 			return $curl;
 		}
@@ -88,7 +86,7 @@
 		 * Adding subscribers to a distribution list
 		 * @param string $list_id
 		 * @param array $data
-		 * @return \stdClass
+		 * @return \CleverMonitor\Api\Connection\Response
 		 */
 		public function putListSubscribers($list_id, array $data) {
 			$curl = $this->connection->put('lists/' . $list_id . '/subscribers', $data);
@@ -99,7 +97,7 @@
 		 * Delete subscribers in a distribution list
 		 * @param string $list_id
 		 * @param array $data
-		 * @return \stdClass
+		 * @return \CleverMonitor\Api\Connection\Response
 		 */
 		public function deleteListSubscribers($list_id, array $data) {
 			$curl = $this->connection->delete('lists/' . $list_id . '/subscribers', $data);
