@@ -12,11 +12,11 @@
 
 		/**
 		 * Get reference
-		 * @param string $name
-		 * @return \stdClass
+		 * @param string $name Name of reference (if NULL, all references will be listed)
+		 * @return \CleverMonitor\Api\Connection\Response
 		 */
-		public function getReference($name) {
-			$curl = $this->connection->get('lists/'.$name);
+		public function getReference($name = NULL) {
+			$curl = $this->connection->get('references/'.$name);
 			return $curl;
 		}
 	}
